requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.9.6
django-preserialize==1.2.1
idna==2.6
MySQL-python==1.2.5
mysqlclient==1.3.12
requests==2.18.4
urllib3==1.22
